requirements.txt
pydantic
uvicorn
qdrant-client
python-dotenv
fastapi
numpy
opencv-python
deepface
tf-keras
torch
redis
slowapi
pyjwt
jwt